requirements.txt
batchgenerators==0.25.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
future==1.0.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
linecache2==1.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
MedPy==0.5.2
monai==1.3.2
monai-generative==0.2.3
mpmath==1.3.0
networkx==3.1
nibabel==5.2.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SimpleITK==2.4.0
six==1.16.0
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2023.7.10
tokenizers==0.20.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
traceback2==1.4.0
transformers==4.46.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
unittest2==1.1.0
urllib3==2.2.3
zipp==3.20.2
